src/secondbrain/voice/voice_enhancement.py
"""
Enhanced voice processing system with advanced features.
"""
import logging
import numpy as np
from typing import Dict, Any, Optional, Tuple
import sounddevice as sd
import librosa
import torch
import torch.nn as nn
from transformers import Wav2Vec2ForCTC, Wav2Vec2Processor
from TTS.api import TTS

logger = logging.getLogger(__name__)

class VoiceEnhancer:
    def __init__(self):
        self.sample_rate = 16000
        self.voice_processor = None
        self.tts_model = None
        self.noise_reducer = None
        self.voice_models = {}
        self.audio_buffer = []
        self.is_initialized = False
        
    async def initialize(self):
        """Initialize voice enhancement components."""
        try:
            # Initialize speech recognition model
            self.voice_processor = Wav2Vec2Processor.from_pretrained("facebook/wav2vec2-base-960h")
            self.model = Wav2Vec2ForCTC.from_pretrained("facebook/wav2vec2-base-960h")
            
            # Initialize text-to-speech
            self.tts_model = TTS("tts_models/en/ljspeech/tacotron2-DDC")
            
            # Initialize noise reduction model
            self.noise_reducer = self._create_noise_reducer()
            
            self.is_initialized = True
            logger.info("Voice enhancement system initialized successfully")
            
        except Exception as e:
            logger.error(f"Error initializing voice enhancement: {e}")
            raise
            
    def _create_noise_reducer(self) -> nn.Module:
        """Create noise reduction model."""
        class NoiseReducer(nn.Module):
            def __init__(self):
                super().__init__()
                self.conv1 = nn.Conv1d(1, 32, 8, 4, 2)
                self.conv2 = nn.Conv1d(32, 32, 8, 4, 2)
                self.conv3 = nn.Conv1d(32, 64, 8, 4, 2)
                self.deconv1 = nn.ConvTranspose1d(64, 32, 8, 4, 2)
                self.deconv2 = nn.ConvTranspose1d(32, 32, 8, 4, 2)
                self.deconv3 = nn.ConvTranspose1d(32, 1, 8, 4, 2)
                self.relu = nn.ReLU()
                
            def forward(self, x):
                x = self.relu(self.conv1(x))
                x = self.relu(self.conv2(x))
                x = self.relu(self.conv3(x))
                x = self.relu(self.deconv1(x))
                x = self.relu(self.deconv2(x))
                x = self.deconv3(x)
                return x
                
        return NoiseReducer()
        
    async def enhance_audio(self, audio_data: np.ndarray) -> Tuple[np.ndarray, Dict[str, Any]]:
        """Enhance audio quality."""
        try:
            # Normalize audio
            audio_normalized = librosa.util.normalize(audio_data)
            
            # Reduce noise
            audio_denoised = await self._reduce_noise(audio_normalized)
            
            # Enhance clarity
            audio_enhanced = await self._enhance_clarity(audio_denoised)
            
            # Calculate quality metrics
            metrics = self._calculate_audio_metrics(audio_enhanced)
            
            return audio_enhanced, metrics
            
        except Exception as e:
            logger.error(f"Error enhancing audio: {e}")
            return audio_data, {}
            
    async def _reduce_noise(self, audio: np.ndarray) -> np.ndarray:
        """Apply noise reduction."""
        try:
            # Convert to tensor
            audio_tensor = torch.FloatTensor(audio).unsqueeze(0).unsqueeze(0)
            
            # Apply noise reduction
            with torch.no_grad():
                denoised = self.noise_reducer(audio_tensor)
                
            return denoised.squeeze().numpy()
            
        except Exception as e:
            logger.error(f"Error reducing noise: {e}")
            return audio
            
    async def _enhance_clarity(self, audio: np.ndarray) -> np.ndarray:
        """Enhance audio clarity."""
        try:
            # Apply basic audio processing
            audio = librosa.effects.preemphasis(audio)
            audio = librosa.effects.trim(audio)[0]
            
            return audio
            
        except Exception as e:
            logger.error(f"Error enhancing clarity: {e}")
            return audio
            
    def _calculate_audio_metrics(self, audio: np.ndarray) -> Dict[str, float]:
        """Calculate audio quality metrics."""
        try:
            metrics = {
                "snr": self._calculate_snr(audio),
                "clarity": self._calculate_clarity(audio),
                "volume": np.abs(audio).mean()
            }
            return metrics
        except Exception as e:
            logger.error(f"Error calculating metrics: {e}")
            return {}
            
    def _calculate_snr(self, audio: np.ndarray) -> float:
        """Calculate signal-to-noise ratio."""
        try:
            signal_power = np.mean(audio ** 2)
            noise = audio - np.mean(audio)
            noise_power = np.mean(noise ** 2)
            return 10 * np.log10(signal_power / noise_power) if noise_power > 0 else 0
        except:
            return 0.0
            
    def _calculate_clarity(self, audio: np.ndarray) -> float:
        """Calculate audio clarity score."""
        try:
            # Simple clarity metric based on spectral centroid
            spec_cent = librosa.feature.spectral_centroid(y=audio, sr=self.sample_rate)
            return float(np.mean(spec_cent))
        except:
            return 0.0
            
    async def enhance_voice_output(self, text: str, voice_params: Dict[str, Any]) -> np.ndarray:
        """Enhance text-to-speech output."""
        try:
            # Generate speech
            audio = self.tts_model.tts(text)
            
            # Apply voice modifications
            audio = self._modify_voice(audio, voice_params)
            
            # Enhance quality
            audio_enhanced, _ = await self.enhance_audio(audio)
            
            return audio_enhanced
            
        except Exception as e:
            logger.error(f"Error enhancing voice output: {e}")
            return None
            
    def _modify_voice(self, audio: np.ndarray, params: Dict[str, Any]) -> np.ndarray:
        """Apply voice modifications."""
        try:
            # Adjust speed
            if "speed" in params:
                audio = librosa.effects.time_stretch(audio, params["speed"])
                
            # Adjust pitch
            if "pitch" in params:
                audio = librosa.effects.pitch_shift(
                    audio,
                    sr=self.sample_rate,
                    n_steps=params["pitch"]
                )
                
            # Adjust volume
            if "volume" in params:
                audio = audio * params["volume"]
                
            return audio
            
        except Exception as e:
            logger.error(f"Error modifying voice: {e}")
            return audio
            
    async def process_stream(self, audio_stream: np.ndarray) -> Tuple[np.ndarray, Dict[str, Any]]:
        """Process real-time audio stream."""
        try:
            # Add to buffer
            self.audio_buffer.extend(audio_stream)
            
            # Process when buffer is large enough
            if len(self.audio_buffer) >= self.sample_rate * 0.5:  # 500ms chunks
                audio_chunk = np.array(self.audio_buffer[:self.sample_rate * 0.5])
                self.audio_buffer = self.audio_buffer[int(self.sample_rate * 0.5):]
                
                # Enhance audio
                enhanced_audio, metrics = await self.enhance_audio(audio_chunk)
                
                return enhanced_audio, metrics
                
            return np.array([]), {}
            
        except Exception as e:
            logger.error(f"Error processing stream: {e}")
            return np.array([]), {}
            
    def get_available_voices(self) -> List[str]:
        """Get list of available voice models."""
        try:
            return self.tts_model.speakers
        except Exception as e:
            logger.error(f"Error getting voices: {e}")
            return []
            
    async def update_voice_model(self, voice_id: str, params: Dict[str, Any]):
        """Update voice model parameters."""
        try:
            if voice_id in self.voice_models:
                self.voice_models[voice_id].update(params)
                logger.info(f"Updated voice model {voice_id}")
        except Exception as e:
            logger.error(f"Error updating voice model: {e}")
            
    def get_voice_quality_metrics(self) -> Dict[str, float]:
        """Get current voice quality metrics."""
        try:
            return {
                "clarity": self._calculate_clarity(np.array(self.audio_buffer)),
                "volume": np.mean(np.abs(self.audio_buffer)) if self.audio_buffer else 0.0,
                "background_noise": self._estimate_noise_level()
            }
        except Exception as e:
            logger.error(f"Error getting quality metrics: {e}")
            return {}
            
    def _estimate_noise_level(self) -> float:
        """Estimate background noise level."""
        try:
            if not self.audio_buffer:
                return 0.0
            # Simple noise estimation using signal variance
            return float(np.var(self.audio_buffer))
        except:
            return 0.0 