pyproject.toml
[tool.poetry]
name = "secondbrain"
version = "0.1.0"
description = "An intelligent personal assistant with voice capabilities"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
pyttsx3 = "2.90"
schedule = "1.2.0"
SpeechRecognition = "3.10.0"
pyaudio = "0.2.13"
python-dotenv = "1.0.0"
tqdm = "4.66.1"
colorama = "0.4.6"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.2"
httpx = "^0.25.2"
psutil = "^5.9.6"
prometheus-client = "^0.19.0"
aiohttp = "^3.9.1"
websockets = "^12.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.23"}
asyncpg = "^0.29.0"
docker = "^7.0.0"
aiodocker = "^0.21.0"
aiosmtplib = "^3.0.1"
prometheus-async = "^22.2.0"
structlog = "^23.2.0"
tenacity = "^8.2.3"
aiodns = "^3.1.1"
cchardet = "^2.1.7"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-cov = "^4.1.0"
black = "23.11.0"
flake8 = "6.1.0"
pre-commit = "^3.5.0"
mypy = "^1.7.0"
pytest-asyncio = "^0.21.1"
pytest-aiohttp = "^1.0.5"
pytest-docker = "^2.0.1"
pytest-mock = "^3.12.0"
pytest-timeout = "^2.2.0"
pytest-xdist = "^3.5.0"
locust = "^2.19.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.poetry.scripts]
secondbrain = "src.secondbrain.main:main" 